pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalbox"
version = "0.1.0"
description = "Script execution control and monitoring"
readme = "README.md"
authors = [
    {name = "pdbeard"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.0.0",
    "PyYAML>=5.4.0",
    "PyQt6>=6.4.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/pdbeard/signalbox"
Repository = "https://github.com/pdbeard/signalbox"
Issues = "https://github.com/pdbeard/signalbox/issues"

[project.scripts]
signalbox = "core.cli_commands:cli"
signalbox-tray = "core.tray_app:main"

[tool.setuptools.package-data]
"core" = ["config/**/*", "icons/*.png"]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0",
    "black>=23.0.0",
    "pytest>=7.0.0",
]

# Development tool scripts
[tool.setuptools]
packages = ["core"]
py-modules = ["signalbox"]

[tool.black]
line-length = 120
target-version = ['py38']

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
per-file-ignores = [
    # Allow f-strings without placeholders in CLI code (used for multiline formatting)
    "core/cli_commands.py:F541",
    "core/executor.py:F541",
    "core/exporters.py:F541",
]
